improved interactive error msgs., use objst stderr instead of NSLog, remove superfluous no stack msg, remove NSLog() of unknown identifiers

+++ Classes/MPWInstanceVariable.h
@@ -12,11 +12,11 @@
 @interface MPWInstanceVariable : MPWObject {
 	NSString	*name;
 	NSString	*type;
-	int			offset;
+	long			offset;
 }
 
 -initWithName:(NSString*)newName offset:(int)newOffset type:(NSString*)newType;
--(int)offset;
+-(long)offset;
 -(NSString*)name;
 -(NSString*)type;
 -valueInContext:anObject;
